+++ src/crawlee/_utils/urls.py
@@ -1,7 +1,12 @@
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 from pydantic import AnyHttpUrl, TypeAdapter
 from yarl import URL
+
+if TYPE_CHECKING:
+    from collections.abc import Iterator
 
 
 def is_url_absolute(url: str) -> bool:
@@ -15,6 +20,15 @@ def is_url_absolute(url: str) -> bool:
 def convert_to_absolute_url(base_url: str, relative_url: str) -> str:
     """Convert a relative URL to an absolute URL using a base URL."""
     return str(URL(base_url).join(URL(relative_url)))
+
+
+def to_absolute_url_iterator(base_url: str, urls: Iterator[str]) -> Iterator[str]:
+    """Convert an iterator of relative URLs to absolute URLs using a base URL."""
+    for url in urls:
+        if is_url_absolute(url):
+            yield url
+        else:
+            yield convert_to_absolute_url(base_url, url)
 
 
 _http_url_adapter = TypeAdapter(AnyHttpUrl)