+++ tests/unit/crawlers/_basic/test_basic_crawler.py
@@ -18,7 +18,7 @@ import pytest
 
 from crawlee import ConcurrencySettings, Glob, service_locator
 from crawlee._request import Request
-from crawlee._types import BasicCrawlingContext, EnqueueLinksKwargs, HttpHeaders
+from crawlee._types import BasicCrawlingContext, EnqueueLinksKwargs, HttpHeaders, HttpMethod
 from crawlee._utils.robots import RobotsTxtFile
 from crawlee.configuration import Configuration
 from crawlee.crawlers import BasicCrawler
@@ -302,34 +302,42 @@ async def test_handles_error_in_failed_request_handler() -> None:
         await crawler.run(['http://a.com/', 'http://b.com/', 'http://c.com/'])
 
 
-async def test_send_request_works(server_url: URL) -> None:
+@pytest.mark.parametrize(
+    ('method', 'path', 'payload'),
+    [
+        pytest.param('GET', 'get', None, id='get send_request'),
+        pytest.param('POST', 'post', b'Hello, world!', id='post send_request'),
+    ],
+)
+async def test_send_request_works(server_url: URL, method: HttpMethod, path: str, payload: None | bytes) -> None:
     response_data: dict[str, Any] = {}
 
     crawler = BasicCrawler(max_request_retries=3)
 
     @crawler.router.default_handler
     async def handler(context: BasicCrawlingContext) -> None:
-        response = await context.send_request(str(server_url))
+        response = await context.send_request(str(server_url / path), method=method, payload=payload)
 
-        response_data['body'] = response.read()
+        response_data['body'] = json.loads(response.read())
         response_data['headers'] = response.headers
 
     await crawler.run(['http://a.com/', 'http://b.com/', 'http://c.com/'])
 
     response_body = response_data.get('body')
     assert response_body is not None
-    assert b'Hello, world!' in response_body
+    assert response_body.get('data') == (payload.decode() if payload else None)
 
     response_headers = response_data.get('headers')
     assert response_headers is not None
     content_type = response_headers.get('content-type')
     assert content_type is not None
-    assert content_type == 'text/html; charset=utf-8'
+    assert content_type == 'application/json'
 
 
 @dataclass
 class AddRequestsTestInput:
     start_url: str
+    loaded_url: str
     requests: Sequence[str | Request]
     expected_urls: Sequence[str]
     kwargs: EnqueueLinksKwargs
@@ -339,6 +347,7 @@ STRATEGY_TEST_URLS = (
     'https://someplace.com/',
     'http://someplace.com/index.html',
     'https://blog.someplace.com/index.html',
+    'https://redirect.someplace.com',
     'https://other.place.com/index.html',
 )
 
@@ -358,6 +367,7 @@ INCLUDE_TEST_URLS = (
         pytest.param(
             AddRequestsTestInput(
                 start_url='https://a.com/',
+                loaded_url='https://a.com/',
                 requests=[
                     'https://a.com/',
                     Request.from_url('http://b.com/'),
@@ -372,52 +382,109 @@ INCLUDE_TEST_URLS = (
         pytest.param(
             AddRequestsTestInput(
                 start_url=STRATEGY_TEST_URLS[0],
+                loaded_url=STRATEGY_TEST_URLS[0],
                 requests=STRATEGY_TEST_URLS,
                 kwargs=EnqueueLinksKwargs(),
                 expected_urls=STRATEGY_TEST_URLS[1:],
             ),
-            id='enqueue_strategy_1',
+            id='enqueue_strategy_default',
         ),
         pytest.param(
             AddRequestsTestInput(
                 start_url=STRATEGY_TEST_URLS[0],
+                loaded_url=STRATEGY_TEST_URLS[0],
                 requests=STRATEGY_TEST_URLS,
                 kwargs=EnqueueLinksKwargs(strategy='all'),
                 expected_urls=STRATEGY_TEST_URLS[1:],
             ),
-            id='enqueue_strategy_2',
+            id='enqueue_strategy_all',
         ),
         pytest.param(
             AddRequestsTestInput(
                 start_url=STRATEGY_TEST_URLS[0],
-                requests=STRATEGY_TEST_URLS,
+                loaded_url=STRATEGY_TEST_URLS[0],
+                requests=STRATEGY_TEST_URLS[:4],
                 kwargs=EnqueueLinksKwargs(strategy='same-domain'),
-                expected_urls=STRATEGY_TEST_URLS[1:3],
+                expected_urls=STRATEGY_TEST_URLS[1:4],
             ),
-            id='enqueue_strategy_3',
+            id='enqueue_strategy_same_domain',
         ),
         pytest.param(
             AddRequestsTestInput(
                 start_url=STRATEGY_TEST_URLS[0],
-                requests=STRATEGY_TEST_URLS,
+                loaded_url=STRATEGY_TEST_URLS[0],
+                requests=STRATEGY_TEST_URLS[:4],
                 kwargs=EnqueueLinksKwargs(strategy='same-hostname'),
                 expected_urls=[STRATEGY_TEST_URLS[1]],
             ),
-            id='enqueue_strategy_4',
+            id='enqueue_strategy_same_hostname',
         ),
         pytest.param(
             AddRequestsTestInput(
                 start_url=STRATEGY_TEST_URLS[0],
+                loaded_url=STRATEGY_TEST_URLS[0],
+                requests=STRATEGY_TEST_URLS[:4],
+                kwargs=EnqueueLinksKwargs(strategy='same-origin'),
+                expected_urls=[],
+            ),
+            id='enqueue_strategy_same_origin',
+        ),
+        # Enqueue strategy with redirect
+        pytest.param(
+            AddRequestsTestInput(
+                start_url=STRATEGY_TEST_URLS[3],
+                loaded_url=STRATEGY_TEST_URLS[0],
+                requests=STRATEGY_TEST_URLS,
+                kwargs=EnqueueLinksKwargs(),
+                expected_urls=STRATEGY_TEST_URLS[:3] + STRATEGY_TEST_URLS[4:],
+            ),
+            id='redirect_enqueue_strategy_default',
+        ),
+        pytest.param(
+            AddRequestsTestInput(
+                start_url=STRATEGY_TEST_URLS[3],
+                loaded_url=STRATEGY_TEST_URLS[0],
+                requests=STRATEGY_TEST_URLS,
+                kwargs=EnqueueLinksKwargs(strategy='all'),
+                expected_urls=STRATEGY_TEST_URLS[:3] + STRATEGY_TEST_URLS[4:],
+            ),
+            id='redirect_enqueue_strategy_all',
+        ),
+        pytest.param(
+            AddRequestsTestInput(
+                start_url=STRATEGY_TEST_URLS[3],
+                loaded_url=STRATEGY_TEST_URLS[0],
+                requests=STRATEGY_TEST_URLS,
+                kwargs=EnqueueLinksKwargs(strategy='same-domain'),
+                expected_urls=STRATEGY_TEST_URLS[:3],
+            ),
+            id='redirect_enqueue_strategy_same_domain',
+        ),
+        pytest.param(
+            AddRequestsTestInput(
+                start_url=STRATEGY_TEST_URLS[3],
+                loaded_url=STRATEGY_TEST_URLS[0],
+                requests=STRATEGY_TEST_URLS,
+                kwargs=EnqueueLinksKwargs(strategy='same-hostname'),
+                expected_urls=[],
+            ),
+            id='redirect_enqueue_strategy_same_hostname',
+        ),
+        pytest.param(
+            AddRequestsTestInput(
+                start_url=STRATEGY_TEST_URLS[3],
+                loaded_url=STRATEGY_TEST_URLS[0],
                 requests=STRATEGY_TEST_URLS,
                 kwargs=EnqueueLinksKwargs(strategy='same-origin'),
                 expected_urls=[],
             ),
-            id='enqueue_strategy_5',
+            id='redirect_enqueue_strategy_same_origin',
         ),
         # Include/exclude
         pytest.param(
             AddRequestsTestInput(
                 start_url=INCLUDE_TEST_URLS[0],
+                loaded_url=INCLUDE_TEST_URLS[0],
                 requests=INCLUDE_TEST_URLS,
                 kwargs=EnqueueLinksKwargs(include=[Glob('https://someplace.com/**/cats')]),
                 expected_urls=[INCLUDE_TEST_URLS[1], INCLUDE_TEST_URLS[4]],
@@ -427,6 +494,7 @@ INCLUDE_TEST_URLS = (
         pytest.param(
             AddRequestsTestInput(
                 start_url=INCLUDE_TEST_URLS[0],
+                loaded_url=INCLUDE_TEST_URLS[0],
                 requests=INCLUDE_TEST_URLS,
                 kwargs=EnqueueLinksKwargs(exclude=[Glob('https://someplace.com/**/cats')]),
                 expected_urls=[INCLUDE_TEST_URLS[2], INCLUDE_TEST_URLS[3]],
@@ -436,6 +504,7 @@ INCLUDE_TEST_URLS = (
         pytest.param(
             AddRequestsTestInput(
                 start_url=INCLUDE_TEST_URLS[0],
+                loaded_url=INCLUDE_TEST_URLS[0],
                 requests=INCLUDE_TEST_URLS,
                 kwargs=EnqueueLinksKwargs(
                     include=[Glob('https://someplace.com/**/cats')], exclude=[Glob('https://**/archive/**')]
@@ -453,6 +522,8 @@ async def test_enqueue_strategy(test_input: AddRequestsTestInput) -> None:
 
     @crawler.router.handler('start')
     async def start_handler(context: BasicCrawlingContext) -> None:
+        # Assign test value to loaded_url - BasicCrawler does not do any navigation by itself
+        context.request.loaded_url = test_input.loaded_url
         await context.add_requests(
             test_input.requests,
             **test_input.kwargs,
@@ -1093,6 +1164,7 @@ async def test_context_use_state_race_condition_in_handlers(key_value_store: Key
     assert (await store.get_value(BasicCrawler._CRAWLEE_STATE_KEY))['counter'] == 2
 
 
+@pytest.mark.run_alone
 @pytest.mark.skipif(sys.version_info[:3] < (3, 11), reason='asyncio.timeout was introduced in Python 3.11.')
 @pytest.mark.parametrize(
     'sleep_type',