+++ tests/unit/http_clients/test_httpx.py
@@ -14,14 +14,18 @@ from crawlee.http_clients import HttpxHttpClient
 from crawlee.statistics import Statistics
 
 if TYPE_CHECKING:
+    from collections.abc import AsyncGenerator
+
     from yarl import URL
 
+    from crawlee.http_clients import HttpClient
     from crawlee.proxy_configuration import ProxyInfo
 
 
 @pytest.fixture
-def http_client() -> HttpxHttpClient:
-    return HttpxHttpClient(http2=False)
+async def http_client() -> AsyncGenerator[HttpClient]:
+    async with HttpxHttpClient(http2=False) as client:
+        yield client
 
 
 async def test_http_1(server_url: URL) -> None:
@@ -128,3 +132,89 @@ async def test_crawl_follow_redirects_false(server_url: URL) -> None:
     assert crawling_result.http_response.status_code == 302
     assert crawling_result.http_response.headers['Location'] == target_url
     assert request.loaded_url == redirect_url
+
+
+async def test_stream(http_client: HttpxHttpClient, server_url: URL) -> None:
+    check_body = b"""\
+<html><head>
+    <title>Hello, world!</title>
+</head>
+<body>
+</body></html>"""
+    content_body: bytes = b''
+
+    async with http_client.stream(str(server_url)) as response:
+        assert response.status_code == 200
+        async for chunk in response.read_stream():
+            content_body += chunk
+
+    assert content_body == check_body
+
+
+async def test_stream_error_double_read_stream(http_client: HttpxHttpClient, server_url: URL) -> None:
+    check_body = b"""\
+<html><head>
+    <title>Hello, world!</title>
+</head>
+<body>
+</body></html>"""
+
+    async with http_client.stream(str(server_url)) as response:
+        assert response.status_code == 200
+        content_body_first: bytes = b''
+        async for chunk in response.read_stream():
+            content_body_first += chunk
+
+        with pytest.raises(RuntimeError):
+            [chunk async for chunk in response.read_stream()]
+
+    assert content_body_first == check_body
+
+
+async def test_stream_error_for_read(http_client: HttpxHttpClient, server_url: URL) -> None:
+    async with http_client.stream(str(server_url)) as response:
+        assert response.status_code == 200
+
+        with pytest.raises(RuntimeError):
+            response.read()
+
+
+async def test_send_request_error_for_read_stream(http_client: HttpxHttpClient, server_url: URL) -> None:
+    response = await http_client.send_request(str(server_url))
+
+    assert response.status_code == 200
+    with pytest.raises(RuntimeError):
+        [item async for item in response.read_stream()]
+
+
+async def test_send_crawl_error_for_read_stream(http_client: HttpxHttpClient, server_url: URL) -> None:
+    response = await http_client.crawl(Request.from_url(str(server_url)))
+    http_response = response.http_response
+
+    assert http_response.status_code == 200
+    with pytest.raises(RuntimeError):
+        [item async for item in http_response.read_stream()]
+
+
+async def test_reuse_context_manager(server_url: URL) -> None:
+    http_client = HttpxHttpClient()
+    async with http_client:
+        response = await http_client.send_request(str(server_url))
+        assert response.status_code == 200
+
+    # Reusing the context manager should not raise an error
+    async with http_client:
+        response = await http_client.send_request(str(server_url))
+        assert response.status_code == 200
+
+
+async def test_work_after_cleanup(http_client: HttpxHttpClient, server_url: URL) -> None:
+    response = await http_client.send_request(str(server_url))
+    assert response.status_code == 200
+
+    # Cleanup the client
+    await http_client.cleanup()
+
+    # After cleanup, the client should still work
+    response = await http_client.send_request(str(server_url))
+    assert response.status_code == 200