+++ website/docusaurus.config.js
@@ -38,6 +38,10 @@ module.exports = {
     githubHost: 'github.com',
     future: {
         experimental_faster: true,
+        v4: {
+            removeLegacyPostBuildHeadAttribute: true,
+            useCssCascadeLayers: false, // this breaks styles on homepage and link colors everywhere
+        },
     },
     headTags: [
         // Intercom messenger